XD-1072 move helper functions to the utils file
